EscSeq* - Adjusted naming, added more sequences, made code more consistent, simplified, etc...
